Move components into correct component folders
