+++ components/Organizer/JudgingTab/allScores.tsx
@@ -1,7 +1,7 @@
 import { Table, Tag, Button, Input, InputRef, Space, TableProps } from 'antd';
 import React, { useRef, useState } from 'react';
 
-import { TeamData, ScoreData, UserData } from '../types/database';
+import { TeamData, ScoreData, UserData } from '../../../types/database';
 import { ExportToCsv } from 'export-to-csv';
 import { FilterConfirmProps, FilterValue, SorterResult } from 'antd/lib/table/interface';
 import { SearchOutlined } from '@ant-design/icons';
