+++ pages/JudgeDash.tsx
@@ -2,7 +2,7 @@ import { Button, Divider, notification, Skeleton } from 'antd';
 import { Dispatch, SetStateAction, useEffect, useState } from 'react';
 import useSWR, { useSWRConfig } from 'swr';
 import JudgingForm from '../components/judges/JudgingForm';
-import { JudgeSchedule } from '../components/schedule';
+import { JudgeSchedule } from '../components/judges/schedule';
 import TeamSelect from '../components/judges/TeamSelect';
 import { JudgingFormFields, ScheduleDisplay, TeamSelectData } from '../types/client';
 import { JudgingSessionData, ResponseError, TeamData } from '../types/database';
